Move the tests to test_machines.py
