+++ inductiva/tests/benchmarks/test_benchmarks.py
@@ -5,7 +5,6 @@ from inductiva._benchmarks.methods import _render_dict
 from inductiva._benchmarks.methods import _compute_tasks_to_run
 from inductiva._benchmarks.methods import _tasks_by_vm_type
 from inductiva.tasks.task import TaskInfo
-from inductiva.resources.machines_base import BaseMachineGroup
 
 executer_4 = mock.MagicMock()
 executer_4.vm_type = "c2-standard-4"
@@ -187,203 +186,3 @@ def test__render_dict_argument_is_object():
     # pylint: disable=W0212
     result = _render_dict(dictionary, argument)
     assert result == {"a": 1, "b": "hello world", "c": 3}
-
-
-def fake_get_quotas_small():
-    """Fake get_quotas function."""
-    return {
-        "max_vcpus": {
-            "max_allowed": 1.0,
-            "in_use": 0.0
-        },
-        "max_instances": {
-            "max_allowed": 1.0,
-            "in_use": 0.0
-        },
-        "max_price_hour": {
-            "max_allowed": 1.0,
-            "in_use": 0.0
-        },
-        "max_disk_size": {
-            "max_allowed": 1.0,
-            "in_use": None
-        },
-        "mg_max_idle": {
-            "max_allowed": 1.0,
-            "in_use": None
-        },
-        "mg_max_ttl": {
-            "max_allowed": 1.0,
-            "in_use": None
-        }
-    }
-
-
-def fake_get_quotas_normal():
-    """Fake get_quotas function."""
-    return {
-        "max_vcpus": {
-            "max_allowed": 100.0,
-            "in_use": 0.0
-        },
-        "max_instances": {
-            "max_allowed": 10.0,
-            "in_use": 0.0
-        },
-        "max_price_hour": {
-            "max_allowed": 20.0,
-            "in_use": 0.0
-        },
-        "max_disk_size": {
-            "max_allowed": 50.0,
-            "in_use": None
-        },
-        "mg_max_idle": {
-            "max_allowed": 150.0,
-            "in_use": None
-        },
-        "mg_max_ttl": {
-            "max_allowed": 48.0,
-            "in_use": None
-        }
-    }
-
-
-def fake_get_quotas_no_total_num_machines():
-    """Fake get_quotas function."""
-    return {
-        "max_vcpus": {
-            "max_allowed": 100.0,
-            "in_use": 0.0
-        },
-        "max_instances": {
-            "max_allowed": 0.0,
-            "in_use": 0.0
-        },
-        "max_price_hour": {
-            "max_allowed": 20.0,
-            "in_use": 0.0
-        },
-        "max_disk_size": {
-            "max_allowed": 50.0,
-            "in_use": None
-        },
-        "mg_max_idle": {
-            "max_allowed": 150.0,
-            "in_use": None
-        },
-        "mg_max_ttl": {
-            "max_allowed": 48.0,
-            "in_use": None
-        }
-    }
-
-
-@mock.patch("inductiva.users.get_quotas", new=fake_get_quotas_small)
-def test_can_start_resource_cant_start_low_quotas():
-    """Test the can_start_resource function when the quotas are small."""
-    resource = mock.MagicMock(spec=BaseMachineGroup)
-    resource.machine_type = "c2-standard-4"
-    resource.estimate_cloud_cost = mock.MagicMock(return_value=0.04724)
-    resource.n_vcpus.total = 4
-    resource.num_machines = 1
-
-    res = BaseMachineGroup.can_start_resource(resource)
-
-    assert not res
-
-
-@mock.patch("inductiva.users.get_quotas", new=fake_get_quotas_normal)
-def test_can_start_resource_cant_start_too_many_vcpu():
-    """Test the can_start_resource function when the resource asks for
-    too many vcpus."""
-
-    resource = mock.MagicMock(spec=BaseMachineGroup)
-    resource.machine_type = "c2-standard-400"
-    resource.estimate_cloud_cost = mock.MagicMock(return_value=0.04724)
-    resource.n_vcpus.total = 400
-    resource.num_machines = 1
-
-    res = BaseMachineGroup.can_start_resource(resource)
-
-    assert not res
-
-
-@mock.patch("inductiva.users.get_quotas",
-            new=fake_get_quotas_no_total_num_machines)
-def test_can_start_resource_cant_start_no_num_machines():
-    """Test the can_start_resource function when the quotas dont allow due to
-    not enough num machines."""
-
-    resource = mock.MagicMock(spec=BaseMachineGroup)
-    resource.machine_type = "c2-standard-4"
-    resource.estimate_cloud_cost = mock.MagicMock(return_value=0.04724)
-    resource.n_vcpus.total = 4
-    resource.num_machines = 1
-
-    res = BaseMachineGroup.can_start_resource(resource)
-
-    assert not res
-
-
-@mock.patch("inductiva.users.get_quotas", new=fake_get_quotas_small)
-def test_can_start_resource_cant_start_high_cost():
-    """Test the can_start_resource function when the resource cost is too
-    high."""
-
-    resource = mock.MagicMock(spec=BaseMachineGroup)
-    resource.machine_type = "c2-standard-1"
-    resource.estimate_cloud_cost = mock.MagicMock(return_value=10.4724)
-    resource.n_vcpus.total = 1
-    resource.num_machines = 1
-
-    res = BaseMachineGroup.can_start_resource(resource)
-
-    assert not res
-
-
-@mock.patch("inductiva.users.get_quotas", new=fake_get_quotas_small)
-def test_can_start_resource_can_start_low_quotas():
-    """Test the can_start_resource function when the quotas are small but can
-    start machine."""
-
-    resource = mock.MagicMock(spec=BaseMachineGroup)
-    resource.machine_type = "c2-standard-1"
-    resource.estimate_cloud_cost = mock.MagicMock(return_value=0.04724)
-    resource.n_vcpus.total = 1
-    resource.num_machines = 1
-
-    res = BaseMachineGroup.can_start_resource(resource)
-
-    assert res
-
-
-@mock.patch("inductiva.users.get_quotas", new=fake_get_quotas_normal)
-def test_can_start_resource_can_start_too_many_vcpu():
-    """Test the can_start_resource function when the resource asks for
-    a valid amount of vcpus."""
-
-    resource = mock.MagicMock(spec=BaseMachineGroup)
-    resource.machine_type = "c2-standard-4"
-    resource.estimate_cloud_cost = mock.MagicMock(return_value=0.04724)
-    resource.n_vcpus.total = 4
-    resource.num_machines = 1
-
-    res = BaseMachineGroup.can_start_resource(resource)
-
-    assert res
-
-
-@mock.patch("inductiva.users.get_quotas", new=fake_get_quotas_small)
-def test_can_start_resource_can_start_high_cost():
-    """Test the can_start_resource function when the resource cost is valid."""
-
-    resource = mock.MagicMock(spec=BaseMachineGroup)
-    resource.machine_type = "c2-standard-1"
-    resource.estimate_cloud_cost = mock.MagicMock(return_value=0.4724)
-    resource.n_vcpus.total = 1
-    resource.num_machines = 1
-
-    res = BaseMachineGroup.can_start_resource(resource)
-
-    assert res
